+++ NEWS.md
@@ -923,6 +923,8 @@ More information at [Dropbox migration guide](https://www.dropbox.com/developers
 * Avoid exporting orphan raster overviews in user migrator
 * Set `soft_geocoding_limit` to default to false.
 * Do not export local visualizations lacking a map
+* Do not export duplicated canonical visualizations
+* Add notifications to user migrator (#13844)
 * Docs, fixed incorrect grammar in en.json file (customer reported).
 
 ### NOTICE